Feat: Delete Degree.css and Job.css
These css refer to these two components, however these components are conditionally rendered within the Education and Experience component pages respectively. It is easier to understand any styles applied to the Degree and Job components in thier larger context with the parent components so their styles will be kept there.